requirements.txt
requests
speechrecognition
pydub 
simpleaudio
google-genai
logging
pydub
flask
